Some code reorganizing
